LOTS OF LOTS OF CHANGES
On a deadline to get this done. just changing a ton of stuff. turned it into a wbsite and notj ust a webpage because the 1 page thing was feeling kind of sort of unbreathable and tight. learning how to work with multiple html files but only keeping 1 js file because that's what internet people said but my project isn't that big so idk  if  its the best move, and I thought about using some frame works so that I could keep the navbar done through some sort of automation but idec at this point, i knoew i could do it in django but that's like taking a "cannon to hunt a rabbit" - mihawk vs zoro @ baratie. good stuff. not just in the live action either. one piece >> caught up to changer 1091 yall go read it. i digress.