got the scroll effect to work- manually
It's hilarious actually. the scrolling effect works manually, so when you scroll, you can do infinitely, the images add perfectly as I wanted them to. but when I try to do it automatticaly it doesn't work. gonna have to revamp it and remake it such taht each image moves automatically...uploading this so i can roll back if i need