I screamed when I figured out the bug.
so stupid- i copied the mobile.js file from an older script.js file and stopped working on it for so long that it was just to do with the simle.html or non-html check i did awhile back. dang dang dang. at least it was an easy fix. thanks responsinator.com